pyproject.toml
[project]
name = "banana-game"
version = "0.1.0"
description = "Fantasy adventure game using Gemini Flash Image"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "google-genai>=0.1.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
packages = ["backend"]